patch auctions

+++ core/src/main/java/net/mizukilab/pit/command/PitAdminSimpleCommand.kt
@@ -573,7 +573,7 @@ class PitAdminSimpleCommand {
     @Execute(name = "StartCustomAuction")
     @Permission("pit.admin")
     @HandHasItem
-    fun startCustomAuction(@Context player: Player, @Arg("price") price: Double) {
+    fun startCustomAuction(@Context player: Player, @Arg("price") price: Long) {
         val itemStack = player.itemInHand
         AuctionEvent().also {
             it.lots = AuctionEvent.LotsData(arrayOf(itemStack.clone()), price, 0)
